Separate "future" connectors methods into separate files with their tests - Updated division logic to always return a float
