+++ apollo-federation/src/sources/connect/json_selection/methods/future/eq.rs
@@ -1,0 +1,109 @@
+use apollo_compiler::collections::IndexMap;
+use serde_json_bytes::Value as JSON;
+use shape::Shape;
+use shape::location::SourceId;
+
+use crate::impl_arrow_method;
+use crate::sources::connect::json_selection::ApplyToError;
+use crate::sources::connect::json_selection::ApplyToInternal;
+use crate::sources::connect::json_selection::MethodArgs;
+use crate::sources::connect::json_selection::PathList;
+use crate::sources::connect::json_selection::VarsWithPathsMap;
+use crate::sources::connect::json_selection::apply_to::ApplyToResultMethods;
+use crate::sources::connect::json_selection::immutable::InputPath;
+use crate::sources::connect::json_selection::location::Ranged;
+use crate::sources::connect::json_selection::location::WithRange;
+
+impl_arrow_method!(EqMethod, eq_method, eq_shape);
+/// Returns true if argument is equal to the applied to value or false if they are not equal.
+/// Simple examples:
+///
+/// $->echo(123)->eq(123)       results in true
+/// $->echo(123)->eq(456)       results in false
+fn eq_method(
+    method_name: &WithRange<String>,
+    method_args: Option<&MethodArgs>,
+    data: &JSON,
+    vars: &VarsWithPathsMap,
+    input_path: &InputPath<JSON>,
+    tail: &WithRange<PathList>,
+) -> (Option<JSON>, Vec<ApplyToError>) {
+    if let Some(MethodArgs { args, .. }) = method_args {
+        if args.len() == 1 {
+            let (value_opt, arg_errors) = args[0].apply_to_path(data, vars, input_path);
+            let matches = if let Some(value) = value_opt {
+                data == &value
+            } else {
+                false
+            };
+            return tail
+                .apply_to_path(&JSON::Bool(matches), vars, input_path)
+                .prepend_errors(arg_errors);
+        }
+    }
+    (
+        None,
+        vec![ApplyToError::new(
+            format!(
+                "Method ->{} requires exactly one argument",
+                method_name.as_ref()
+            ),
+            input_path.to_vec(),
+            method_name.range(),
+        )],
+    )
+}
+#[allow(dead_code)] // method type-checking disabled until we add name resolution
+fn eq_shape(
+    method_name: &WithRange<String>,
+    _method_args: Option<&MethodArgs>,
+    _input_shape: Shape,
+    _dollar_shape: Shape,
+    _named_var_shapes: &IndexMap<&str, Shape>,
+    source_id: &SourceId,
+) -> Shape {
+    Shape::bool(method_name.shape_location(source_id))
+}
+
+#[cfg(test)]
+mod tests {
+    use serde_json_bytes::json;
+
+    use crate::selection;
+
+    #[test]
+    fn eq_should_return_true_when_applied_to_equals_argument() {
+        assert_eq!(
+            selection!(
+                r#"
+                    result: value->eq(123)
+                "#
+            )
+            .apply_to(&json!({ "value": 123 })),
+            (
+                Some(json!({
+                    "result": true,
+                })),
+                vec![],
+            ),
+        );
+    }
+
+    #[test]
+    fn eq_should_return_false_when_applied_to_does_not_equal_argument() {
+        assert_eq!(
+            selection!(
+                r#"
+                    result: value->eq(1234)
+                "#
+            )
+            .apply_to(&json!({ "value": 123 })),
+            (
+                Some(json!({
+                    "result": false,
+                })),
+                vec![],
+            ),
+        );
+    }
+}
