+++ apollo-federation/src/sources/connect/json_selection/methods/public/first.rs
@@ -15,10 +15,12 @@ use crate::sources::connect::json_selection::location::Ranged;
 use crate::sources::connect::json_selection::location::WithRange;
 
 impl_arrow_method!(FirstMethod, first_method, first_shape);
-/// The "first" method is a utility function that can be run against an array to grab the 0th item from it.
+/// The "first" method is a utility function that can be run against an array to grab the 0th item from it
+/// or a string to get the first character.
 /// The simplest possible example:
 ///
 /// $->echo([1,2,3])->first     results in 1
+/// $->echo("hello")->first     results in "h"
 fn first_method(
     method_name: &WithRange<String>,
     method_args: Option<&MethodArgs>,
@@ -120,5 +122,13 @@ mod tests {
     #[test]
     fn first_should_get_none_when_no_items_exist() {
         assert_eq!(selection!("$->first").apply_to(&json!([])), (None, vec![]),);
+    }
+
+    #[test]
+    fn first_should_get_first_char_from_string() {
+        assert_eq!(
+            selection!("$->first").apply_to(&json!("hello")),
+            (Some(json!("h")), vec![]),
+        );
     }
 }
