+++ apollo-federation/src/sources/connect/json_selection/methods/public/size.rs
@@ -16,10 +16,12 @@ use crate::sources::connect::json_selection::location::Ranged;
 use crate::sources::connect::json_selection::location::WithRange;
 
 impl_arrow_method!(SizeMethod, size_method, size_shape);
-/// Returns the number of items in an array.
+/// Returns the number of items in an array, length of a string, or the number of properties in an array.
 /// The simplest possible example:
 ///
-/// $->echo([1,2,3,4,5])->size     would result in 5
+/// $->echo([1,2,3,4,5])->size                      would result in 5
+/// $->echo("hello")->size                          would result in 5
+/// $->echo({"a": true, "b": true})->size           would result in 2
 fn size_method(
     method_name: &WithRange<String>,
     method_args: Option<&MethodArgs>,
@@ -192,6 +194,38 @@ mod tests {
                     "range": [7, 11],
                 }))]
             ),
+        );
+    }
+
+    #[test]
+    fn size_should_return_length_of_string() {
+        assert_eq!(
+            selection!("$->size").apply_to(&json!("hello")),
+            (Some(json!(5)), vec![]),
+        );
+    }
+
+    #[test]
+    fn size_should_return_0_on_empty_string() {
+        assert_eq!(
+            selection!("$->size").apply_to(&json!("")),
+            (Some(json!(0)), vec![]),
+        );
+    }
+
+    #[test]
+    fn size_should_return_number_of_properties_of_an_object() {
+        assert_eq!(
+            selection!("$->size").apply_to(&json!({ "a": 1, "b": 2, "c": 3 })),
+            (Some(json!(3)), vec![]),
+        );
+    }
+
+    #[test]
+    fn size_should_return_0_on_empty_object() {
+        assert_eq!(
+            selection!("$->size").apply_to(&json!({})),
+            (Some(json!(0)), vec![]),
         );
     }
 }
